+++ client_test.go
@@ -1,6 +1,7 @@
 package inout
 
 import (
+	"context"
 	"io/ioutil"
 	"testing"
 
@@ -8,9 +9,8 @@ import (
 )
 
 func Test_fetch(t *testing.T) {
-	ts := startServer(clientHTML, 200)
-	defer ts.Close()
-	res := fetch(ts.URL)
+	defer stopServer(context.Background(), startServer(clientHTML, 200))
+	res := fetch(source)
 
 	assert.NotNil(t, res)
 
@@ -26,7 +26,7 @@ func Test_fetch(t *testing.T) {
 func Test_fetch_retries(t *testing.T) {
 	ts := startServer("", 429)
 	defer ts.Close()
-	res := fetch(ts.URL)
+	res := fetch(source)
 	assert.NotNil(t, res)
 	assert.NotNil(t, res.err)
 	assert.Equal(t, 3, res.attempts)