+++ headless.go
@@ -6,31 +6,75 @@ import (
 	"strings"
 
 	"github.com/chromedp/cdproto/cdp"
+	"github.com/chromedp/cdproto/dom"
 	"github.com/chromedp/chromedp"
 )
 
-func waitForDomElement(ctx context.Context, selector, addr string, verbose bool) (string, error) {
+type headlesResult struct {
+	htmlDoc  string
+	imgBytes []byte
+}
+
+func headless(ctx context.Context, c *config) (*headlesResult, error) {
 	// create context
 	childCtx, cancel := chromedp.NewContext(ctx)
 	defer cancel()
 
-	// navigate
-	if err := chromedp.Run(childCtx, chromedp.Navigate(addr)); err != nil {
-		return "", fmt.Errorf("could not navigate to %s: %v", addr, err)
+	var res strings.Builder
+	var img []byte
+	if err := chromedp.Run(childCtx, chromeTasks(c, &res, 90, &img)); err != nil {
+		err = fmt.Errorf("error in running headless to %s: %w", c.source, err)
+		return nil, err
 	}
 
-	var nodes []*cdp.Node
-	err := chromedp.Run(childCtx, chromedp.Nodes(selector, &nodes, chromedp.ByQueryAll, chromedp.NodeVisible))
-	if err != nil {
-		return "", fmt.Errorf("could not get nodes %s: %v", selector, err)
-	}
+	return &headlesResult{htmlDoc: strings.TrimSpace(res.String()), imgBytes: img}, nil
+}
 
-	var sb strings.Builder
-	for _, n := range nodes {
-		var cnt string
-		if err := chromedp.Run(childCtx, chromedp.OuterHTML(n.FullXPath(), &cnt)); err == nil {
-			sb.WriteString(cnt)
-		}
+// chromeTasks ...
+func chromeTasks(c *config, res *strings.Builder, quality int, buf *[]byte) chromedp.Tasks {
+	tasks := []chromedp.Action{}
+	if c.screenshot {
+		tasks = append(tasks, chromedp.EmulateViewport(1920, 2000))
 	}
-	return strings.TrimSpace(sb.String()), nil
+	tasks = append(tasks, chromedp.Navigate(c.source))
+	if c.waitUntil > 0 {
+		tasks = append(tasks, chromedp.Sleep(c.waitUntil))
+	}
+	if len(c.waitFor) > 0 {
+		tasks = append(tasks, chromedp.WaitReady(c.waitFor))
+	}
+	if len(c.query) > 0 {
+		var nodes []*cdp.Node
+		tasks = append(tasks,
+			chromedp.WaitReady(c.query),
+			chromedp.Nodes(c.query, &nodes),
+			chromedp.ActionFunc(func(c context.Context) error {
+				for _, v := range nodes {
+					str, err := dom.GetOuterHTML().WithNodeID(v.NodeID).Do(c)
+					if err != nil {
+						return err
+					}
+					res.WriteString(str)
+				}
+				return nil
+			}),
+		)
+	} else {
+		tasks = append(tasks, chromedp.ActionFunc(func(c context.Context) error {
+			node, err := dom.GetDocument().Do(c)
+			if err != nil {
+				return err
+			}
+			str, err := dom.GetOuterHTML().WithNodeID(node.NodeID).Do(c)
+			if err != nil {
+				return err
+			}
+			res.WriteString(str)
+			return nil
+		}))
+	}
+	if c.screenshot {
+		tasks = append(tasks, chromedp.CaptureScreenshot(buf))
+	}
+	return tasks
 }